broke down analytics into small components, with old api
